TODO (LAB 5)/TODO/main/serializers.py
from rest_framework import serializers
from main.models import Lists, Tasks


class ListSerializer(serializers.ModelSerializer):
    tasks = serializers.StringRelatedField(many=True, read_only=True)

    class Meta:
        model = Lists
        fields = '__all__'


class ListDetailSerializer(serializers.ModelSerializer):
    tasks = serializers.StringRelatedField(many=True, read_only=True)

    class Meta:
        model = Lists
        fields = ('name', 'tasks', )


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tasks
        fields = ('id', 'task', 'mark', 'lists_id')


class TaskDetailSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tasks
        fields = '__all__'

