TODO (LAB 5)/TODO/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-crispy-forms==1.11.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1

djangorestframework~=3.12.2
djangorestframework-jwt~=1.11.0
